+++ packages/ui-code-editor/docs/changelog.md
@@ -2,6 +2,10 @@
 title: Changelog
 ---
 
+## NEXT
+
+-   Updated `@codemirror` packages to `^6.0.0`
+
 ## 1.0.0
 
 -   Initial release